+++ samples/unity-of-bugs/Assets/Scripts/SceneButtons.cs
@@ -6,6 +6,7 @@ public class SceneButtons : MonoBehaviour
     public void LoadBugFarm() => SceneManager.LoadScene("1_BugFarm");
     public void LoadNativeSupport() => SceneManager.LoadScene("2_NativeSupport");
     public void LoadAdditionalSamples() => SceneManager.LoadScene("3_AdditionalSamples");
+    public void LoadThreadedSamples() => SceneManager.LoadScene("4_ThreadedSamples");
 
     public void CloseGame()
     {